Modificada classe InfoGym

+++ app/src/main/java/com/example/sandra/gymapp/classesjava/InfoGym.java
@@ -1,5 +1,6 @@
 package com.example.sandra.gymapp.classesjava;
 
+
 /**
  * Created by 47419119l on 28/04/16.
  */
@@ -8,8 +9,8 @@ public class InfoGym {
     String direccionGym;
     int telefonoGym;
     String correoElectronicoGym;
-    long latitudGym;
-    long longitudGym;
+    double latitudGym;
+    double longitudGym;
     String horarioGym[];
 
     public String getNombreGym() {
@@ -44,19 +45,19 @@ public class InfoGym {
         this.correoElectronicoGym = correoElectronicoGym;
     }
 
-    public long getLatitudGym() {
+    public double getLatitudGym() {
         return latitudGym;
     }
 
-    public void setLatitudGym(long latitudGym) {
+    public void setLatitudGym(double latitudGym) {
         this.latitudGym = latitudGym;
     }
 
-    public long getLongitudGym() {
+    public double getLongitudGym() {
         return longitudGym;
     }
 
-    public void setLongitudGym(long longitudGym) {
+    public void setLongitudGym(double longitudGym) {
         this.longitudGym = longitudGym;
     }
